requirements.txt
pynput==1.7.6
pygetwindow==0.0.9
pyautogui==0.9.53
pyscreeze==0.1.29
Pillow==10.4.0
pywin32
mss
opencv-python
numpy